+++ crates/nostr-relay-pool/src/relay/stats.rs
@@ -17,6 +17,9 @@ use nostr::Timestamp;
 #[cfg(not(target_arch = "wasm32"))]
 use tokio::sync::RwLock;
 
+#[cfg(not(target_arch = "wasm32"))]
+use super::constants::LATENCY_MAX_VALUES;
+
 /// Ping Stats
 #[cfg(not(target_arch = "wasm32"))]
 #[derive(Debug, Clone)]
@@ -165,6 +168,13 @@ impl RelayConnectionStats {
         sum.checked_div(latencies.len() as u32)
     }
 
+    /// Get number of latency reads
+    #[cfg(not(target_arch = "wasm32"))]
+    pub(crate) async fn latency_reads(&self) -> usize {
+        let latencies = self.latencies.read().await;
+        latencies.len()
+    }
+
     pub(crate) fn new_attempt(&self) {
         self.attempts.fetch_add(1, Ordering::SeqCst);
     }
@@ -194,7 +204,7 @@ impl RelayConnectionStats {
     #[cfg(not(target_arch = "wasm32"))]
     pub(crate) async fn save_latency(&self, latency: Duration) {
         let mut latencies = self.latencies.write().await;
-        if latencies.len() >= 5 {
+        if latencies.len() >= LATENCY_MAX_VALUES {
             latencies.pop_back();
         }
         latencies.push_front(latency)